+++ src/main/resources/db/migration/V1.7__post.sql
@@ -1,8 +1,9 @@
 create table post(
-    id serial NOT NULL PRIMARY KEY,
-    "time" timestamp NOT NULL,
-    author_id int4 NOT NULL REFERENCES person (id),
-    title text,
-    post_text text,
-    is_blocked int2,
-    is_deleted int2);
+                     id         bigserial NOT NULL PRIMARY KEY,
+                     "time"     timestamp NOT NULL,
+                     author_id  bigint    NOT NULL REFERENCES person (id),
+                     title      text,
+                     post_text  text,
+                     is_blocked int2,
+                     is_deleted int2
+);
