+++ src/main/resources/db/migration/V1.9__post_like.sql
@@ -1,5 +1,7 @@
-create table post_like(
-    id serial NOT NULL PRIMARY KEY,
-    "time" timestamp,
-    person_id int4 NOT NULL REFERENCES person (id),
-    post_id int4 NOT NULL REFERENCES post (id));
+create table post_like
+(
+    id        bigserial NOT NULL PRIMARY KEY,
+    "time"    timestamp,
+    person_id bigint    NOT NULL REFERENCES person (id),
+    post_id   bigint    NOT NULL REFERENCES post (id)
+);
